reorganize & dedupe first-time-flow selectors
