+++ ui/app/pages/first-time-flow/metametrics-opt-in/metametrics-opt-in.container.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 import MetaMetricsOptIn from './metametrics-opt-in.component'
 import { setParticipateInMetaMetrics } from '../../../store/actions'
-import { getFirstTimeFlowTypeRoute } from '../first-time-flow.selectors'
+import { getFirstTimeFlowTypeRoute } from '../../../selectors'
 
 const firstTimeFlowTypeNameMap = {
   create: 'Selected Create New Wallet',
