+++ ui/app/pages/first-time-flow/select-action/select-action.container.js
@@ -2,7 +2,7 @@ import { connect } from 'react-redux'
 import { withRouter } from 'react-router-dom'
 import { compose } from 'redux'
 import { setFirstTimeFlowType } from '../../../store/actions'
-import { getFirstTimeFlowTypeRoute } from '../first-time-flow.selectors'
+import { getFirstTimeFlowTypeRoute } from '../../../selectors'
 import Welcome from './select-action.component'
 
 const mapStateToProps = (state) => {
