+++ ui/app/selectors/index.js
@@ -1,5 +1,6 @@
 export * from './confirm-transaction'
 export * from './custom-gas'
+export * from './first-time-flow'
 export * from './permissions'
 export * from './selectors'
 export * from './tokens'
